add new questions

+++ 1878A.cpp
@@ -1,10 +1,10 @@
 #include<iostream>
 using namespace std;
 int main(){
-    int t;
+    long long t;
     cin>>t;
     while(t--){
-        int n,k;
+        long long n,k;
         cin>>n>>k;
         int flag=0;
         for(int i=0;i<n;i++){
