+++ 1878C.cpp
@@ -1,10 +1,10 @@
 #include<bits/stdc++.h>
 using namespace std;
 int main(){
-    int t;
+    long long t;
     cin>>t;
     while(t--){
-        int n,k,x;
+        long long n,k,x;
         cin>>n>>k>>x;
         long long front=(1LL*k*(k+1))/2;
         long long back=(1LL*k*(n+n+1-k))/2;
